requirements.txt
# PRISM POC - Python 패키지 의존성
# Local sLLM (Ollama) 버전

# 핵심 프레임워크
streamlit==1.32.0

# PDF 처리
PyPDF2==3.0.1
pdf2image==1.17.0
pdfplumber==0.11.0
Pillow==10.4.0

# 이미지 처리
opencv-python-headless==4.9.0.80

# 데이터 처리
numpy==1.26.4
pandas==2.2.1

# 품질 평가
nltk==3.8.1
rouge-score==0.1.2
sentence-transformers==2.5.1

# HTTP 클라이언트 (Ollama API)
requests==2.31.0

# 환경 변수
python-dotenv==1.0.1

# 테스트
pytest==8.1.1
pytest-cov==4.1.0